requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0

# Natural Language Processing
textblob==0.17.1
spacy==3.7.2
scikit-learn==1.4.0
networkx==3.2.1
wordcloud==1.9.3

# Web Development
flask==2.0.1
flask-restful==0.3.10
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
apispec==6.3.0
apispec-webframeworks==0.5.2

# Report Generation
reportlab==4.0.8
python-pptx==0.6.23
python-docx==1.0.1

# Data Processing and Visualization
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2

# Performance and Scalability
redis==4.5.4
celery==5.2.7
pymongo==4.3.3
motor==3.3.2
aiohttp==3.9.3
uvicorn==0.27.1
gunicorn==20.1.0

# Testing and Monitoring
pytest==7.4.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
prometheus-client==0.19.0
sentry-sdk==1.40.4

# Integration
simple-salesforce==1.12.5
hubspot-api-client==8.2.0
zendesk-api==0.1.0

# Development Tools
black==23.7.0
flake8==6.1.0
mypy==1.5.1
isort==5.12.0

nltk==3.8.1
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.0
python-pptx==0.6.23
python-docx==1.0.1
matplotlib==3.8.2
seaborn==0.13.0
flask-restful==0.3.10
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
apispec==6.3.0
apispec-webframeworks==0.5.2
hubspot-api-client==8.2.0
zendesk-api==0.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl 